requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.3=py39h5eee18b_0
aiosignal=1.2.0=pyhd8ed1ab_0
astor=0.8.1=py39h06a4308_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=openblas
blinker=1.5=pyhd8ed1ab_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bs4=4.11.1=hd8ed1ab_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
cryptography=38.0.1=py39h9ce1e76_0
cudatoolkit=11.1.1=h6406543_8
cudnn=8.1.0.77=h90431f1_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
fftw=3.3.9=h27cfd23_1
flatbuffers=2.0.0=h2531618_0
fontconfig=2.13.1=hef1e5e3_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.3=py39h5eee18b_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h36c2ea0_2
glib=2.69.1=h4ff587b_1
google-auth=2.14.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py39hec9cf62_0
hdf5=1.10.6=nompi_h6a2412b_1114
icu=58.2=hf484d3e_1000
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.0.0=pyha770c72_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h166bdaf_1
keras=2.4.3=hd3eb1b0_0
keras-base=2.4.3=pyhd3eb1b0_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py39h295c915_0
krb5=1.19.3=h3790be6_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=15_linux64_openblas
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=15_linux64_openblas
libclang=10.0.1=default_hb85057a_2
libcurl=7.85.0=h91b91d3_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h2818925_1
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.5.2=py39h06a4308_0
matplotlib-base=3.5.2=py39hf590b9c_0
multidict=6.0.2=py39h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.19.2=py39hc896f84_1
numpy-base=1.19.2=py39h21a3de8_1
oauthlib=3.2.2=pyhd8ed1ab_0
openssl=1.1.1s=h7f8727e_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pillow=9.2.0=py39hace64e9_1
pip=22.2.2=py39h06a4308_0
ply=3.11=py39h06a4308_0
protobuf=3.20.1=py39h295c915_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pysocks=1.7.1=pyha2e5f31_6
python=3.9.15=haa1d7c7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39h7f8727e_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.3=py39h6a678d5_0
scipy=1.7.3=py39hf838250_2
setuptools=65.5.0=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=hbd366e4_1
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.39.3=h5082296_0
tensorboard=2.9.0=py39h06a4308_0
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.4.1=mkl_py39h4683426_0
tensorflow-addons=0.13.0=py39_1
tensorflow-base=2.4.1=mkl_py39h43e0292_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
tensorflow-gpu=2.5.0=py39_cuda11.1_cudnn8.1_1
termcolor=2.1.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py39h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
typeguard=2.13.0=py_1
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.11=pyhd8ed1ab_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=py39h5eee18b_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.8.1=py39h5eee18b_0
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
